indexing/pyproject.toml
[project]
name = "code-embedding"
version = "0.1.0"
description = "Simple example for cocoindex: build embedding index based on source code."
requires-python = ">=3.11"
dependencies = [
    "cocoindex[embeddings]>=0.2.8",
    "python-dotenv>=1.0.1",
    "pgvector>=0.4.1",
    "psycopg[binary,pool]",
]

[tool.setuptools]
packages = []